+++ windows/runner/process_control_channel.h
@@ -1,0 +1,8 @@
+#ifndef RUNNER_PROCESS_CONTROL_CHANNEL_H_
+#define RUNNER_PROCESS_CONTROL_CHANNEL_H_
+
+#include <flutter/binary_messenger.h>
+
+void RegisterProcessControlChannel(flutter::BinaryMessenger* messenger);
+
+#endif  // RUNNER_PROCESS_CONTROL_CHANNEL_H_